improvements to checksum
Now stores checksum directly after the firmware image + the xor of the checksum in the following word, both have to be valid to pass.
"../config.h" includes in assembler startup changed to "config.h" to allow values to be overridden by a user supplied configuration, this means that no changes need to be made to the sboot_stm32 tree.
Firmware crypt updated to new checksum method
Added slow crc32, consumes less flash but is slower than the table lookup method.